Finished the perceptron

+++ Neural Network Tools/src/tools/Classifier.java
@@ -3,6 +3,6 @@ package tools;
 @FunctionalInterface
 public interface Classifier {
 
-	public int classify(float x, float y);
+	public float classify(float x, float y);
 	
 }
